+++ CMake/BaseCode.cmake
@@ -85,9 +85,10 @@ function(ttk_add_base_template_library library)
   if(uppercase_CMAKE_BUILD_TYPE MATCHES RELEASE)
     if(TTK_ENABLE_CPU_OPTIMIZATION AND NOT MSVC)
       if (CMAKE_OSX_ARCHITECTURES MATCHES arm64) # Apple Silicon
-        target_compile_options(${library} INTERFACE -mcpu=apple-m1 -O3)
+        target_compile_options(${library} INTERFACE -mcpu=apple-m1)
       else()
-        target_compile_options(${library} INTERFACE -march=native -O3)
+        # -O3 already enabled by CMake's Release configuration
+        target_compile_options(${library} INTERFACE -march=native)
       endif()
     endif()
   endif()