+++ core/base/common/FlatJaggedArray.h
@@ -29,6 +29,14 @@ namespace ttk {
       this->offsets_ = std::move(offsets);
     }
 
+    /**
+     * @brief Clear the underlaying vectors
+     */
+    inline void clear() {
+      this->data_.clear();
+      this->offsets_.clear();
+    }
+
     // ############################## //
     // Mimic the vector of vector API //
     // ############################## //
@@ -55,6 +63,61 @@ namespace ttk {
       }
 #endif
       return this->offsets_[id];
+    }
+
+    struct Slice {
+      const SimplexId *const ptr;
+      const SimplexId len;
+      inline SimplexId operator[](const SimplexId id) {
+#ifndef TTK_ENABLE_KAMIKAZE
+        if(id < 0 || id >= this->len) {
+          return -1;
+        }
+#endif
+        return this->ptr[id];
+      }
+      inline const SimplexId *data() const {
+        return this->ptr;
+      }
+      inline const SimplexId *begin() const {
+        return this->ptr;
+      }
+      inline const SimplexId *end() const {
+        return this->ptr + this->len;
+      }
+      inline SimplexId size() const {
+        return this->len;
+      }
+    };
+
+    struct Iterator {
+      size_t id;
+      const FlatJaggedArray &parent;
+      inline void operator++() {
+        this->id++;
+      }
+      inline Slice operator*() const {
+        return this->parent[this->id];
+      }
+      inline bool operator!=(const Iterator &other) const {
+        return this->id != other.id;
+      }
+    };
+
+    inline Iterator begin() const {
+      return {0, *this};
+    }
+    inline Iterator end() const {
+      return {this->size(), *this};
+    }
+
+    inline Slice operator[](const SimplexId id) const {
+#ifndef TTK_ENABLE_KAMIKAZE
+      if(id < 0 || id >= (SimplexId)this->offsets_.size()) {
+        return {nullptr, 0};
+      }
+#endif
+      return {this->get_ptr(id, 0), this->size(id)};
     }
 
     /**
@@ -97,7 +160,10 @@ namespace ttk {
     /**
      * @brief Returns the number of sub-vectors
      */
-    inline size_t subvectorsNumber() const {
+    inline size_t size() const {
+      if(this->empty()) {
+        return 0;
+      }
       return this->offsets_.size() - 1;
     }
 
@@ -154,16 +220,16 @@ namespace ttk {
      */
     void copyTo(std::vector<std::vector<SimplexId>> &dst,
                 int threadNumber = 1) const {
-      dst.resize(this->subvectorsNumber());
-      for(size_t i = 0; i < this->subvectorsNumber(); ++i) {
-        dst[i].resize(this->size(i));
+      dst.resize(this->size());
+      for(size_t i = 0; i < this->size(); ++i) {
+        dst[i].resize((*this)[i].size());
       }
 #ifdef TTK_ENABLE_OPENMP
 #pragma omp parallel for num_threads(threadNumber)
 #endif // TTK_ENABLE_OPENMP
-      for(size_t i = 0; i < this->subvectorsNumber(); ++i) {
+      for(size_t i = 0; i < this->size(); ++i) {
         for(size_t j = 0; j < dst[i].size(); ++j) {
-          dst[i][j] = this->get(i, j);
+          dst[i][j] = (*this)[i][j];
         }
       }
       TTK_FORCE_USE(threadNumber);
@@ -178,9 +244,9 @@ namespace ttk {
      */
     inline void writeToFile(const std::string &fName) const {
       std::ofstream out(fName);
-      for(size_t i = 0; i < this->subvectorsNumber(); ++i) {
-        for(SimplexId j = 0; j < this->size(i); ++j) {
-          out << this->get(i, j) << " ";
+      for(const auto &slice : *this) {
+        for(const auto el : slice) {
+          out << el << " ";
         }
         out << '\n';
       }