+++ core/base/localizedTopologicalSimplification/LocalizedTopologicalSimplification.h
@@ -164,7 +164,7 @@ namespace ttk {
           const IT &vOrder = order[v];
           const IT nNeighbors = triangulation->getVertexNeighborNumber(v);
           for(IT n = 0; n < nNeighbors; n++) {
-            IT u;
+            IT u{-1};
             triangulation->getVertexNeighbor(v, n, u);
             if(vOrder < order[u]) {
               hasLargerNeighbor = true;
@@ -264,7 +264,7 @@ namespace ttk {
           IT numberOfLargerNeighbors = 0;
           IT numberOfLargerNeighborsThisThreadVisited = 0;
           for(IT n = 0; n < nNeighbors; n++) {
-            IT u;
+            IT u{-1};
             triangulation->getVertexNeighbor(v, n, u);
 
             const IT &orderU = order[u];
@@ -308,7 +308,7 @@ namespace ttk {
             std::vector<Propagation<IT> *> neighborPropagations(
               nNeighbors, nullptr);
             for(IT n = 0; n < nNeighbors; n++) {
-              IT u;
+              IT u{-1};
               triangulation->getVertexNeighbor(v, n, u);
               if(propagationMask[u] != nullptr) {
                 auto &neighborPropagation = neighborPropagations[n];
@@ -398,7 +398,7 @@ namespace ttk {
           IT numberOfLargerNeighbors = 0;
           IT numberOfLargerNeighborsThisThreadVisited = 0;
           for(IT n = 0; n < nNeighbors; n++) {
-            IT u;
+            IT u{-1};
             triangulation->getVertexNeighbor(v, n, u);
 
             const IT &orderU = order[u];
@@ -442,7 +442,7 @@ namespace ttk {
             std::vector<Propagation<IT> *> neighborPropagations(
               nNeighbors, nullptr);
             for(IT n = 0; n < nNeighbors; n++) {
-              IT u;
+              IT u{-1};
               triangulation->getVertexNeighbor(v, n, u);
               if(propagationMask[u] != nullptr) {
                 auto &neighborPropagation = neighborPropagations[n];
@@ -637,7 +637,7 @@ namespace ttk {
 
             IT nNeighbors = triangulation->getVertexNeighborNumber(v);
             for(IT n = 0; n < nNeighbors; n++) {
-              IT u;
+              IT u{-1};
               triangulation->getVertexNeighbor(v, n, u);
               auto &s = segmentation[u];
               if(s >= 0 && order[u] > saddleOrder) {
@@ -771,7 +771,7 @@ namespace ttk {
 
           IT nNeighbors = triangulation->getVertexNeighborNumber(v);
           for(IT n = 0; n < nNeighbors; n++) {
-            IT u;
+            IT u{-1};
             triangulation->getVertexNeighbor(v, n, u);
 
             // if u is in segment and has not already been added to the queue
@@ -858,7 +858,7 @@ namespace ttk {
 
             IT nNeighbors = triangulation->getVertexNeighborNumber(v);
             for(IT n = 0; n < nNeighbors; n++) {
-              IT u;
+              IT u{-1};
               triangulation->getVertexNeighbor(v, n, u);
 
               if(u == saddleIndex) {