+++ core/vtk/ttkPersistenceDiagram/ttkPersistenceDiagram.h
@@ -37,7 +37,7 @@
 /// Herbert Edelsbrunner and John Harer \n
 /// American Mathematical Society, 2010
 ///
-/// Four backends are available for the computation:
+/// Five backends are available for the computation:
 ///
 ///  1) FTM \n
 /// \b Related \b publication \n
@@ -51,16 +51,24 @@
 /// Jules Vidal, Pierre Guillou, Julien Tierny\n
 /// IEEE Transactions on Visualization and Computer Graphics, 2021
 ///
-/// 3) Persistent Simplex \n
-/// This is a textbook (and very slow) algorithm, described in
-/// "Algorithm and Theory of Computation Handbook (Second Edition)
-/// - Special Topics and Techniques" by Atallah and Blanton on page 97.
+/// 3) Discrete Morse Sandwich (default) \n
+/// \b Related \b publication \n
+/// "Discrete Morse Sandwich: Fast Computation of Persistence Diagrams for
+/// Scalar Data -- An Algorithm and A Benchmark" \n
+/// Pierre Guillou, Jules Vidal, Julien Tierny \n
+/// Technical Report, arXiv:2206.13932, 2022 \n
+/// Fast and versatile algorithm for persistence diagram computation.
 ///
 /// 4) Approximate Approach \n
 /// \b Related \b publication \n
 /// "Fast Approximation of Persistence Diagrams with Guarantees" \n
 /// Jules Vidal, Julien Tierny\n
-/// IEEE Symposium on Large Data Visualization and Analysis (LDAV), 2021\n
+/// IEEE Symposium on Large Data Visualization and Analysis (LDAV), 2021
+///
+/// 5) Persistent Simplex \n
+/// This is a textbook (and very slow) algorithm, described in
+/// "Algorithm and Theory of Computation Handbook (Second Edition)
+/// - Special Topics and Techniques" by Atallah and Blanton on page 97.
 ///
 /// \sa ttkFTMTreePP
 /// \sa ttkPersistenceCurve
@@ -167,6 +175,9 @@ public:
 
   vtkGetMacro(Epsilon, double);
   vtkSetMacro(Epsilon, double);
+
+  vtkSetMacro(IgnoreBoundary, bool);
+  vtkGetMacro(IgnoreBoundary, bool);
 
 protected:
   ttkPersistenceDiagram();